requirements.txt
requests
python-dotenv
colorama==0.4.6